Replace BigDecimal with price in cents
* Spring Frebase does not support BigDecimal for Java 17

+++ src/main/java/com/example/splitbill/ReceiptLine.java
@@ -2,13 +2,13 @@ package com.example.splitbill;
 
 import java.math.BigDecimal;
 
-record ReceiptLine(String title, BigDecimal price) {
+record ReceiptLine(String title, Integer price) {
 
     public String getTitle() {
         return title;
     }
 
-    public BigDecimal getPrice() {
+    public Integer getPrice() {
         return price;
     }
 }
